# openapi/product.json
{
  "components": {
    "schemas": {
...
      "CommandRunCancelReceipt": {
        "additionalProperties": false,
        "properties": {
          "command_run": {
            "$ref": "#/components/schemas/CommandRunView"
          },
          "receipt_id": {
            "title": "Receipt Id",
            "type": "string"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          }
        },
        "required": [
          "receipt_id",
          "status_message",
          "command_run"
        ],
        "title": "CommandRunCancelReceipt",
        "type": "object"
      },
      "CommandRunCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          }
        },
        "required": [
          "action_id",
          "confirmed"
        ],
        "title": "CommandRunCancelRequest",
        "type": "object"
      },
      "CommandRunOutputPage": {
        "additionalProperties": false,
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "is_bounded": {
            "title": "Is Bounded",
            "type": "boolean"
          },
          "is_changed": {
            "title": "Is Changed",
            "type": "boolean"
          },
          "is_missing": {
            "title": "Is Missing",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "output_complete": {
            "title": "Output Complete",
            "type": "boolean"
          }
        },
        "required": [
          "command_id",
          "content",
          "output_complete",
          "is_missing",
          "is_changed",
          "is_bounded"
        ],
        "title": "CommandRunOutputPage",
        "type": "object"
      },
      "CommandRunProductState": {
        "enum": [
          "queued",
          "running",
          "cancelling",
          "succeeded",
          "failed",
          "timed_out",
          "cancelled"
        ],
        "type": "string"
      },
      "CommandRunView": {
        "additionalProperties": false,
        "properties": {
          "cancel_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskMemberReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Summary"
          },
          "output_complete": {
            "title": "Output Complete",
            "type": "boolean"
          },
          "output_href": {
            "title": "Output Href",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/CommandRunProductState"
          }
        },
        "required": [
          "id",
          "purpose",
          "state",
          "created_at",
          "output_href",
          "output_complete"
        ],
        "title": "CommandRunView",
        "type": "object"
      },
      "FileReference": {
        "additionalProperties": false,
        "description": "An ordered navigation value to one ordinary workspace file.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "FileReference",
        "type": "object"
      },
...
      "HumanRequestItem": {
        "additionalProperties": false,
        "properties": {
          "allow_other": {
            "default": false,
            "title": "Allow Other",
            "type": "boolean"
          },
          "allow_skip": {
            "default": false,
            "title": "Allow Skip",
            "type": "boolean"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/HumanRequestOption"
                },
                "maxItems": 3,
                "minItems": 2,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "response_schema": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Schema"
          }
        },
        "required": [
          "id",
          "prompt"
        ],
        "title": "HumanRequestItem",
        "type": "object"
      },
...
      "HumanRequestProductStatus": {
        "enum": [
          "open",
          "answered",
          "expired",
          "cancelled"
        ],
        "type": "string"
      },
      "HumanRequestResolutionView": {
        "additionalProperties": false,
        "properties": {
          "resolved_at": {
            "format": "date-time",
            "title": "Resolved At",
            "type": "string"
          },
          "status": {
            "enum": [
              "answered",
              "expired",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "status",
          "summary",
          "resolved_at"
        ],
        "title": "HumanRequestResolutionView",
        "type": "object"
      },
...
      "HumanRequestResponseReceipt": {
        "additionalProperties": false,
        "properties": {
          "continuation_pending": {
            "const": true,
            "default": true,
            "title": "Continuation Pending",
            "type": "boolean"
          },
          "receipt_id": {
            "title": "Receipt Id",
            "type": "string"
          },
          "request": {
            "$ref": "#/components/schemas/HumanRequestView"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          }
        },
        "required": [
          "receipt_id",
          "status_message",
          "request"
        ],
        "title": "HumanRequestResponseReceipt",
        "type": "object"
      },
      "HumanRequestResponseRequest": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "input": {
            "$ref": "#/components/schemas/HumanRequestResponseInput"
          }
        },
        "required": [
          "action_id",
          "input"
        ],
        "title": "HumanRequestResponseRequest",
        "type": "object"
      },
...
      "HumanRequestView": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "cancel_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FileReference"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HumanRequestItem"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "enum": [
              "input",
              "direction",
              "approval",
              "review"
            ],
            "title": "Kind",
            "type": "string"
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskMemberReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanRequestResolutionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/HumanRequestProductStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "summary",
          "items",
          "opened_at",
          "status"
        ],
        "title": "HumanRequestView",
        "type": "object"
      },
      "JsonValue": {},
...
      "OperationFailure": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProductFailureCode"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "suggested_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Next Step"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "code",
          "summary",
          "retryable"
        ],
        "title": "OperationFailure",
        "type": "object"
      },
      "ProductAction": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "$ref": "#/components/schemas/ProductActionConfirmation"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_schema": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Schema"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "label",
          "href",
          "confirmation"
        ],
        "title": "ProductAction",
        "type": "object"
      },
...
      "ProductFailureCode": {
        "description": "Closed, product-safe failure vocabulary shared by HTTP and Operator.",
        "enum": [
          "invalid_request",
          "not_found",
          "conflict",
          "cursor_reset_required",
          "access_denied",
          "unavailable",
          "internal_error"
        ],
        "title": "ProductFailureCode",
        "type": "string"
      },
...
      "TaskMemberReference": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TaskMemberReference",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Banksia Product API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tasks/{task_id}/command-runs/{command_id}": {
      "get": {
        "operationId": "get_command_run_tasks__task_id__command_runs__command_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Command Run",
        "tags": [
          "command-runs"
        ]
      }
    },
    "/tasks/{task_id}/command-runs/{command_id}/cancel": {
      "post": {
        "operationId": "post_command_run_cancel_tasks__task_id__command_runs__command_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRunCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRunCancelReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Command Run Cancel",
        "tags": [
          "command-runs"
        ]
      }
    },
    "/tasks/{task_id}/command-runs/{command_id}/output": {
      "get": {
        "operationId": "get_command_run_output_tasks__task_id__command_runs__command_id__output_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 65536,
              "maximum": 65536,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandRunOutputPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Command Run Output",
        "tags": [
          "command-runs"
        ]
      }
    },
...
    "/tasks/{task_id}/human-requests/{request_id}": {
      "get": {
        "operationId": "get_human_request_tasks__task_id__human_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanRequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Human Request",
        "tags": [
          "human-requests"
        ]
      }
    },
    "/tasks/{task_id}/human-requests/{request_id}/responses": {
      "post": {
        "operationId": "post_human_request_response_tasks__task_id__human_requests__request_id__responses_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanRequestResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanRequestResponseReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Human Request Response",
        "tags": [
          "human-requests"
        ]
      }
    },
...
  }
}